fix: example이 없을 때 db에 insert하지 않도록 수정

+++ server/src/problems/problems.service.ts
@@ -105,25 +105,26 @@ export class ProblemsService {
     await this.dataSource.manager.transaction(
       async (transactionEntityManager: EntityManager) => {
         await transactionEntityManager.save(problem);
+        if (updateProblemDTO.examples) {
+          await transactionEntityManager
+            .createQueryBuilder()
+            .delete()
+            .from(Example)
+            .where('problemId = :id', { id: id })
+            .execute();
 
-        await transactionEntityManager
-          .createQueryBuilder()
-          .delete()
-          .from(Example)
-          .where('problemId = :id', { id: id })
-          .execute();
-
-        await transactionEntityManager
-          .createQueryBuilder()
-          .insert()
-          .into(Example)
-          .values(
-            [...updateProblemDTO.examples].map((example) => ({
-              ...example,
-              problemId: id,
-            })),
-          )
-          .execute();
+          await transactionEntityManager
+            .createQueryBuilder()
+            .insert()
+            .into(Example)
+            .values(
+              updateProblemDTO.examples.map((example) => ({
+                ...example,
+                problemId: id,
+              })),
+            )
+            .execute();
+        }
       },
     );
   }
